chore: Refactor File Stucture
migrate from single `lib.rs` to:
- instructions
- state
- constants
lib.rs
errors.rs
