+++ programs/airdrop-armor/src/constants/seeds.rs
@@ -1,0 +1,3 @@
+pub const PROGRAM_CONFIG_SEED: [u8; 14] = *b"program_config";
+
+pub const ARMOR_SEED: [u8; 5] = *b"armor";
