+++ programs/airdrop-armor/src/errors.rs
@@ -1,0 +1,7 @@
+use anchor_lang::error_code;
+
+#[error_code]
+pub enum ArmorError {
+    #[msg("Claim Wallet does not match armor.claim_wallet")]
+    Unauthorized,
+}
