+++ programs/airdrop-armor/src/instructions/mod.rs
@@ -1,0 +1,14 @@
+pub mod initialize;
+pub use initialize::*;
+
+pub mod create_armor;
+pub use create_armor::*;
+
+pub mod verify_armor;
+pub use verify_armor::*;
+
+pub mod update_claim_wallet;
+pub use update_claim_wallet::*;
+
+pub mod destroy_armor;
+pub use destroy_armor::*;
